refactor: extracting share responsibility and creating useTerritoryActionShare
